CSSTUDIO-2072 Move the context menu-items for copying and appending PV names to the clipboard from 'org.phoebus.applications.probe' to 'org.csstudio.display.builder.runtime'.
