+++ app/display/runtime/src/main/java/org/csstudio/display/builder/runtime/ContextMenuCopyPvToClipboardWithDescription.java
@@ -1,4 +1,4 @@
-package org.phoebus.applications.probe;
+package org.csstudio.display.builder.runtime;
 
 import javafx.scene.image.Image;
 import javafx.scene.input.Clipboard;
