+++ app/display/runtime/src/main/java/org/csstudio/display/builder/runtime/Messages.java
@@ -14,7 +14,13 @@ public class Messages
 {
     // Keep in alphabetical order and aligned with messages.properties
     /** Localized message */
-    public static String NavigateBack_TT,
+    public static String AppendPVNameToClipboard,
+                         AppendPVNameToClipboardWithDescription,
+                         Copy,
+                         CopyPVNameToClipboardWithDescription,
+                         CopySubMenu,
+                         CopyWithValue,
+                         NavigateBack_TT,
                          NavigateForward_TT,
                          OpenDataBrowser,
                          OpenInEditor,
