+++ app/display/runtime/src/main/resources/META-INF/services/org.phoebus.ui.spi.ContextMenuEntry
@@ -1,1 +1,2 @@
+org.csstudio.display.builder.runtime.ContextMenuCopySubMenu
 org.csstudio.display.builder.runtime.app.ProbeDisplayContextMenuEntry
