+++ app/display/runtime/src/main/resources/org/csstudio/display/builder/runtime/messages.properties
@@ -1,3 +1,9 @@
+AppendPVNameToClipboard=Append PV Name to Clipboard
+AppendPVNameToClipboardWithDescription=Append PV Name to Clipboard with Description...
+Copy=Copy PV Name to Clipboard
+CopyPVNameToClipboardWithDescription=Copy PV Name to Clipboard with Description...
+CopySubMenu=Copy PV Name...
+CopyWithValue=Copy PV Name to Clipboard with Value
 NavigateBack_TT=Open previous display
 NavigateForward_TT=Open next display
 OpenDataBrowser=Open Data Browser
