+++ app/probe/src/main/resources/META-INF/services/org.phoebus.ui.spi.ContextMenuEntry
@@ -1,2 +1,1 @@
-org.phoebus.applications.probe.ContextMenuCopySubMenu
 org.phoebus.applications.probe.ContextLaunchProbe
