+++ app/probe/src/main/resources/org/phoebus/applications/probe/messages.properties
@@ -1,12 +1,6 @@
 Alarm=Alarm:
 Alarms=Alarms: 
-AppendPVNameToClipboard=Append PV Name to Clipboard
-AppendPVNameToClipboardWithDescription=Append PV Name to Clipboard with Description...
 ControlRange=Control Range: 
-Copy=Copy PV Name to Clipboard
-CopyPVNameToClipboardWithDescription=Copy PV Name to Clipboard with Description...
-CopySubMenu=Copy PV Name...
-CopyWithValue=Copy PV Name to Clipboard with Value
 DisplayRange=Display Range: 
 EnumLbls=Enumeration Labels:
 Format=Format:
